requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.2.7
Django==4.1.6
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-faker==0.2
django-filter==22.1
django-import-export==3.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.4
et-xmlfile==1.1.0
fake-factory==9999.9.9
Faker==16.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
Pillow==9.3.0
psycopg2==2.9.5
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
